+++ src/components/nav-drawer/nav-drawer.spec.ts
@@ -149,46 +149,6 @@ describe('Navigation Drawer', () => {
       computedStyles = window.getComputedStyle(mini);
       expect(computedStyles.getPropertyValue('display')).to.not.equal('none');
     });
-
-    it('should emit events on drawer toggle', async () => {
-      const eventSpy = sinon.spy(el, 'emitEvent');
-      el.toggle();
-
-      expect(eventSpy).calledWith('igcOpening');
-      expect(eventSpy).calledWith('igcOpened');
-
-      el.toggle();
-
-      expect(eventSpy).calledWith('igcClosing');
-      expect(eventSpy).calledWith('igcClosed');
-    });
-
-    it('should successfully cancel opening event', async () => {
-      el.addEventListener('igcOpening', (ev: CustomEvent) => {
-        ev.preventDefault();
-      });
-
-      const eventSpy = sinon.spy(el, 'emitEvent');
-
-      el.show();
-
-      await elementUpdated(el);
-      expect(eventSpy).calledOnceWith('igcOpening');
-    });
-
-    it('should successfully cancel closing events', async () => {
-      el.show();
-      el.addEventListener('igcClosing', (ev: CustomEvent) => {
-        ev.preventDefault();
-      });
-
-      const eventSpy = sinon.spy(el, 'emitEvent');
-
-      el.hide();
-
-      await elementUpdated(el);
-      expect(eventSpy).calledOnceWith('igcClosing');
-    });
   });
 
   const createNavDrawerElement = (