+++ src/components/nav-drawer/nav-drawer.ts
@@ -5,13 +5,6 @@ import { EventEmitterMixin } from '../common/mixins/event-emitter';
 import { partNameMap } from '../common/util';
 import { styles } from './nav-drawer.material.css';
 
-export interface IgcNavDrawerEventMap {
-  igcOpening: CustomEvent<any>;
-  igcOpened: CustomEvent<void>;
-  igcClosing: CustomEvent<any>;
-  igcClosed: CustomEvent<void>;
-}
-
 /**
  * Represents a side navigation container that provides
  * quick access between views.
@@ -21,19 +14,11 @@ export interface IgcNavDrawerEventMap {
  * @slot - The default slot for the drawer.
  * @slot mini - The slot for the mini variant of the drawer.
  *
- * @fires igcOpening - Emitted when the navigation drawer is about to open.
- * @fires igcOpened - Emitted when the navigation drawer is opened.
- * @fires igcClosing - Emitted when the navigation drawer is about to close.
- * @fires igcClosed - Emitted when the navigation drawer is closed.
- *
  * @csspart base - The base wrapper of the navigation drawer.
  * @csspart main - The main container.
  * @csspart mini - The mini container.
  */
-export default class IgcNavDrawerComponent extends EventEmitterMixin<
-  IgcNavDrawerEventMap,
-  Constructor<LitElement>
->(LitElement) {
+export default class IgcNavDrawerComponent extends LitElement {
   /** @private */
   public static tagName = 'igc-nav-drawer';
 
@@ -54,12 +39,7 @@ export default class IgcNavDrawerComponent extends EventEmitterMixin<
       return;
     }
 
-    if (!this.handleOpening()) {
-      return;
-    }
-
     this.open = true;
-    this.emitEvent('igcOpened');
   }
 
   /** Closes the drawer. */
@@ -68,12 +48,7 @@ export default class IgcNavDrawerComponent extends EventEmitterMixin<
       return;
     }
 
-    if (!this.handleClosing()) {
-      return;
-    }
-
     this.open = false;
-    this.emitEvent('igcClosed');
   }
 
   /** Toggles the open state of the drawer. */
@@ -83,22 +58,6 @@ export default class IgcNavDrawerComponent extends EventEmitterMixin<
     } else {
       this.show();
     }
-  }
-
-  private handleOpening() {
-    const args = {
-      cancelable: true,
-    };
-
-    return this.emitEvent('igcOpening', args);
-  }
-
-  private handleClosing() {
-    const args = {
-      cancelable: true,
-    };
-
-    return this.emitEvent('igcClosing', args);
   }
 
   private resolvePartNames(base: string) {